+++ services/thumbnails/pkg/thumbnail/thumbnail.go
@@ -3,7 +3,6 @@ package thumbnail
 import (
 	"bytes"
 	"image"
-	"image/gif"
 	"mime"
 
 	"github.com/owncloud/ocis/v2/ocis-pkg/log"
@@ -70,15 +69,12 @@ type SimpleManager struct {
 // Generate creates a thumbnail and stores it
 func (s SimpleManager) Generate(r Request, img interface{}) (string, error) {
 	var match image.Rectangle
-	var inputDimensions image.Rectangle
-	switch m := img.(type) {
-	case *gif.GIF:
-		match = s.resolutions.ClosestMatch(r.Resolution, m.Image[0].Bounds())
-		inputDimensions = m.Image[0].Bounds()
-	case image.Image:
-		match = s.resolutions.ClosestMatch(r.Resolution, m.Bounds())
-		inputDimensions = m.Bounds()
+
+	inputDimensions, err := r.Generator.Dimensions(img)
+	if err != nil {
+		return "", err
 	}
+	match = s.resolutions.ClosestMatch(r.Resolution, inputDimensions)
 
 	// validate max input image dimensions - 6016x4000
 	if inputDimensions.Size().X > s.maxDimension.X || inputDimensions.Size().Y > s.maxDimension.Y {
